requirements.txt
Flask
python-dotenv

selenium